Make the reader generic also
