+++ build.sbt
@@ -32,7 +32,7 @@ val simulacrumVersion = "0.12.0"
 val sparkVersion = "2.3.0"
 val tensorflowVersion = "1.8.0"
 val xgBoostVersion = "0.72-20180627-1214081f"
-val shapelessDatatypeVersion = "0.1.6"
+val shapelessDatatypeVersion = "0.1.9"
 
 val CompileTime = config("compile-time").hide
