+++ core/src/main/scala/com/spotify/featran/transformers/Identity.scala
@@ -17,6 +17,8 @@
 
 package com.spotify.featran.transformers
 
+import com.spotify.featran.FlatReader
+
 /**
  * Transform features by passing them through.
  *
