+++ core/src/main/scala/com/spotify/featran/transformers/NHotWeightedEncoder.scala
@@ -17,7 +17,7 @@
 
 package com.spotify.featran.transformers
 
-import com.spotify.featran.{FeatureBuilder, FeatureRejection}
+import com.spotify.featran.{FeatureBuilder, FeatureRejection, FlatReader}
 
 import scala.collection.SortedMap
 import scala.collection.mutable.{Map => MMap, Set => MSet}
